fix(transaction): widen transaction ids to 64-bit and read them as doubles
Transaction ids come from `DBDescriptor::transactionGetNextId()` and were a
`uint32_t` counter surfaced to JS via `napi_create_uint32`, but read back in
`TransactionLog::AddEntry` with `napi_get_value_int32`. That applies JS ToInt32
semantics: it wraps modulo 2^32, so every id at or above 2^31 arrived negative
and was rejected as "transaction id must be a non-negative integer".

Once a database's counter passed ~2.1 billion transactions, every subsequent
`log.addEntry()` threw, for the life of the process — the counter is in-memory,
so only a restart cleared it. Records and secondary indices still committed
(Harper's `updateIndices()` runs before `writeCommit()`), so the process kept
serving while silently writing no audit/transaction-log entries at all: no
replication out, no replay in.

The same ToInt32 wrap also aliased out-of-range ids onto live ones rather than
rejecting them, so an entry could bind to the wrong transaction.

The id is a live-registry key only (`DBDescriptor::transactionGet`) and is never
persisted in the log frame, so widening is internal and changes no on-disk
format:

- `nextTransactionId`, `TransactionHandle::id`, the registry map key,
  `TransactionLogHandle::transactionId` and `addEntry()` are now `uint64_t`.
- `Transaction::Id`, `Transaction::UseLog` and the registry diagnostics dump
  emit doubles (exact to 2^53) instead of `napi_create_uint32`.
- `readTransactionId()` parses as a double and validates integrality and range
  before narrowing, rejecting anything above `Number.MAX_SAFE_INTEGER` so two
  ids can never collide in the registry.

Headroom goes from 2^31 to 2^53. At the ~6,500 transactions/sec observed on a
busy database that is 28 years of continuous process uptime instead of 3.8 days.

+++ src/binding/database/db_descriptor.cpp
@@ -1674,7 +1674,7 @@ void DBDescriptor::transactionAdd(std::shared_ptr<TransactionHandle> txnHandle) 
 /**
  * Retrieves a transaction from the registry.
  */
-std::shared_ptr<TransactionHandle> DBDescriptor::transactionGet(uint32_t id) {
+std::shared_ptr<TransactionHandle> DBDescriptor::transactionGet(uint64_t id) {
 	std::lock_guard<std::mutex> lock(this->txnsMutex);
 	auto it = this->transactions.find(id);
 	if (it != this->transactions.end()) {
@@ -1696,7 +1696,7 @@ void DBDescriptor::transactionRemove(std::shared_ptr<TransactionHandle> txnHandl
 	auto it = this->transactions.find(txnHandle->id);
 	if (it != this->transactions.end()) {
 		if (it->second != txnHandle) {
-			DEBUG_LOG("%p DBDescriptor::transactionRemove txnId %u mismatch! expected %p, got %p\n", this, txnHandle->id, it->second.get(), txnHandle.get());
+			DEBUG_LOG("%p DBDescriptor::transactionRemove txnId %llu mismatch! expected %p, got %p\n", this, (unsigned long long)txnHandle->id, it->second.get(), txnHandle.get());
 		}
 		this->transactions.erase(it);
 	}
@@ -1723,7 +1723,7 @@ void DBDescriptor::closeTransactionsByEnv(napi_env env) {
 	}
 
 	for (auto& txnHandle : toClose) {
-		DEBUG_LOG("%p DBDescriptor::closeTransactionsByEnv closing transaction %u (env=%p)\n", this, txnHandle->id, env);
+		DEBUG_LOG("%p DBDescriptor::closeTransactionsByEnv closing transaction %llu (env=%p)\n", this, (unsigned long long)txnHandle->id, env);
 		txnHandle->close();
 		// close() can only self-remove while it can still reach this descriptor
 		// through its DBHandle, and a handle closed earlier by the user has
@@ -1738,7 +1738,7 @@ void DBDescriptor::closeTransactionsByEnv(napi_env env) {
 /**
  * Generates the next unique transaction ID for this database.
  */
-uint32_t DBDescriptor::transactionGetNextId() {
+uint64_t DBDescriptor::transactionGetNextId() {
 	return ++this->nextTransactionId;
 }
